// 7-wonders-calculator/src/app/modules/points-calculator/interfaces/pointsCalculationInput.ts
export interface PointsCalculationInput {
  playerName: string;
  money: number;
  military: number;
  wonders: number;
  blueCards: number;
  yellowCards: number;
  citiesCards: number;
  leadersCards: number;
  edificeCards: number;
  armadaCards: number;
  guilds: number;
  other: number;
  scienceGearCount: number;
  scienceCompassCount: number;
  sciencePlateCount: number;
}